게시글 삭제시 권한확인

+++ src/main/java/com/example/prj1be20231109/controller/BoardController.java
@@ -50,7 +50,16 @@ public class BoardController {
     }
 
     @DeleteMapping("remove/{id}")
-    public ResponseEntity remove(@PathVariable Integer id) {
+    public ResponseEntity remove(@PathVariable Integer id,
+                                 @SessionAttribute(value = "login", required = false) Member login) {
+        if (login == null) {
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build(); // 401
+        }
+//        니가 누군지 알지만 권한이 없다.
+        if (!service.hasAccess(id, login)) {
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).build(); // 403
+        }
+
         if (service.remove(id)) {
             //잘됬으면 이거
             return ResponseEntity.ok().build();
